+++ dejacode/settings.py
@@ -29,7 +29,7 @@ environ.Env.read_env(ENV_FILE)  # Reading the .env file into os.environ
 SECRET_KEY = env.str("SECRET_KEY")
 ALLOWED_HOSTS = env.list(
     "ALLOWED_HOSTS",
-    default=[".localhost", "127.0.0.1", "[::1]", "host.docker.internal"],
+    default=[".localhost", "127.0.0.1", "[::1]", "host.docker.internal", "172.17.0.1"],
 )
 CSRF_TRUSTED_ORIGINS = env.list("CSRF_TRUSTED_ORIGINS", default=[])
 # SECURITY WARNING: don't run with debug turned on in production
@@ -489,7 +489,7 @@ LOGGING = {
         "dejacode_toolkit": {
             "handlers": ["null"] if IS_TESTS else ["console"],
             "propagate": False,
-            "level": DEJACODE_LOG_LEVEL,
+            "level": "DEBUG" if DEBUG else DEJACODE_LOG_LEVEL,
         },
         "django_auth_ldap": {
             "handlers": ["null"] if IS_TESTS else ["console"],