+++ CHANGELOG.rst
@@ -3,6 +3,12 @@ Release notes
 
 ### Version 5.0.2-dev
 
+- Lookup in PurlDB by purl in Add Package form.
+  When a Package URL is available in the context of the "Add Package" form,
+  for example when using a link from the Vulnerabilities tab,
+  data is fetched from the PurlDB to initialize the form.
+  https://github.com/nexB/dejacode/issues/47
+
 ### Version 5.0.1
 
 - Improve the stability of the "Check for new Package versions" feature.