+++ CHANGELOG.rst
@@ -1,10 +1,21 @@
 Release notes
 =============
 
-### Version 5.0.1-dev
+### Version 5.0.2-dev
+
+- Lookup in PurlDB by purl in Add Package form.
+  When a Package URL is available in the context of the "Add Package" form,
+  for example when using a link from the Vulnerabilities tab,
+  data is fetched from the PurlDB to initialize the form.
+  https://github.com/nexB/dejacode/issues/47
+
+### Version 5.0.1
 
 - Improve the stability of the "Check for new Package versions" feature.
   https://github.com/nexB/dejacode/issues/17
+
+- Improve the support for SourgeForge download URLs.
+  https://github.com/nexB/dejacode/issues/26
 
 ### Version 5.0.0